Implement Vue Router and separate maintenance page
The changes introduce Vue Router to the application and move the
maintenance page content to a separate component. This improves the
overall structure and allows for future expansion with additional
routes. The main App component is simplified, delegating the maintenance
page rendering to the new Maintenance view.

Key changes:
- Added Vue Router configuration
- Created a new Maintenance.vue component
- Simplified App.vue by removing maintenance-specific code
- Updated imports and component structure
- Improved dark mode toggle implementation

These changes lay the groundwork for a more scalable application
structure while maintaining existing functionality.
